first complete version of the basic games
